requirements.txt
altair==5.5.0
annotated-types==0.7.0
arxiv==2.2.0
attrs==25.3.0
bertopic==0.16.4
blinker==1.9.0
blis==1.2.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
cymem==2.0.11
faiss-cpu==1.8.0
feedparser==6.0.12
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
hdbscan==0.8.40
hf-xet==1.1.10
huggingface-hub==0.34.5
idna==3.10
Jinja2==3.1.4
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.5.0
networkx==3.3
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
packaging==24.2
pandas==2.2.2
pillow==10.4.0
plotly==6.3.0
preshed==3.0.10
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.16.2
sentence-transformers==2.7.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.37.1
sympy==1.13.3
tenacity==8.5.0
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.2.2+cpu
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.5.0
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
